OMS-N: installd: add command 'rmidmap'
Add an installd command to remove an idmap file. This is the inverse of
the 'idmap' command and is intended for clean-up once an idmap file is
no longer needed because an APK was removed, etc.

This commit depends on a corresponding commit in frameworks/base (with
the same Change-Id).

Change-Id: Ib6f6ddd1e8031d7e643e72bb8b8a0417e80e17b4